--- habit-tracker-frontend/pages/ResetPassword.tsx
@@ -1,37 +1,0 @@
-import { useSearchParams } from "react-router";
-import { useState } from "react";
-
-function ResetPassword() {
-  const [params] = useSearchParams();
-  const token = params.get("token");
-  const [password, setPassword] = useState("");
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const res = await fetch("http://localhost:3000/api/users/reset-password", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ token, password }),
-    });
-
-    const data = await res.json();
-    alert(data.message || data.error);
-  };
-
-  return (
-    <div>
-      <h2>Restablecer Contraseña</h2>
-      <form onSubmit={handleSubmit}>
-        <input
-          type="password"
-          placeholder="Nueva contraseña"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <button type="submit">Restablecer</button>
-      </form>
-    </div>
-  );
-}
-
-export default ResetPassword;
